src/requirements.txt
apriltag==0.0.16
tensorrt
pycuda
numpy==1.17.4
ffmpeg-python
# opencv-contrib-python==4.4.*
